V.0.5 -> Improved the info page
I've added a text that shows up if it's opened or closed the POI (Points of Interest). I think I've almost done the 80% of the page so cheers!!

Also made that when u click Google Maps it searches for the place and not the coords. (Not ideal for big franchises, though.)

Also fixed the bug when u scroll in the ListView.

(I've focused on Foursquare, I need to build the IPA to see if Yelp or Sostenible works as well)

Fix:
✓ Make Apple Maps only available for Apple
✓ Fix the social media error Icons and make it like Facebook (It's easy but I want to sleep)
✓ Fix the align of the photos
✓ Fix the price don't showing up (It was a type error — what a mess!)"

TODO (From the last commit):
- Yelp info (If Yelp works)
- Allow the user to expand a photo and check if I can see opinions or anything
- Change the API KEYS
- Error 404 if anything goes wrong (ERROR HANDLING)